Moved MXF seek methods to standalone class Added open incomplete header test
